+++ public/release-notes.json
@@ -1,4 +1,4 @@
 {
   "version": "0.1.0",
-  "notes": "enhance: improve LocalStorage adapter with production-ready features"
+  "notes": "feat: complete Storage Factory implementation (Step 0A.4) - Phase 0 COMPLETE!"
 }
