+++ public/sw.js
@@ -73,4 +73,4 @@ self.addEventListener('fetch', (event) => {
     event.respondWith(fetch(request));
   }
 });
-// Build Timestamp: 2025-09-20T07:39:42.799Z
+// Build Timestamp: 2025-09-20T09:37:51.991Z
